extract registerUser from register_form, debug response handlers in registerUser
