+++ src/components/register/register_form.js
@@ -1,31 +1,7 @@
 import React, { Component } from 'react';
 import { View, TextInput, StyleSheet, TouchableOpacity, Text, KeyboardAvoidingView } from 'react-native';
 import styles from '../../../src/styles/register/registerFormStyles'
-
-const registerUser = (email, name, password, passwordConfirmation) => {
-  return fetch('https://vast-wildwood-58678.herokuapp.com/users', {
-    method: 'POST',
-    headers: {
-      Accept: 'application/json',
-      'Content-Type': 'application/json'
-    },
-    body: JSON.stringify({
-      user:{
-        name: name,
-        email: email,
-        password: password,
-        password_confirmation: passwordConfirmation
-      }
-    })
-  })
-  .then((response) => response.json())
-  .then((responseJson) => {
-    alert(responseJson.name)
-  })
-  .catch((error) => {
-    alert(error)
-  })
-}
+import registerUser from '../../../src/fetch/createUser'
 
 export default class RegisterForm extends Component {
   constructor(props){
